Small refactors to event table
